+++ poem/src/lib.rs
@@ -74,6 +74,8 @@ pub use middleware::Middleware;
 pub use poem_derive::handler;
 pub use request::{Request, RequestBuilder, RequestParts};
 pub use response::{Response, ResponseBuilder, ResponseParts};
-pub use route::{connect, delete, get, head, options, patch, post, put, trace, Route, RouteMethod};
+pub use route::{
+    connect, delete, get, head, options, patch, post, put, trace, Route, RouteDomain, RouteMethod,
+};
 pub use server::Server;
 pub use web::{FromRequest, IntoResponse, RequestBody};