+++ poem/src/route/mod.rs
@@ -1,10 +1,16 @@
 //! Route object and DSL
 
+mod internal;
 mod router;
-mod tree;
+mod router_domain;
+mod router_method;
 
+pub(crate) use internal::radix_tree::PathParams;
 #[allow(unreachable_pub)]
-pub use router::{
-    connect, delete, get, head, options, patch, post, put, trace, Route, RouteMethod,
+pub use router::Route;
+#[allow(unreachable_pub)]
+pub use router_domain::RouteDomain;
+#[allow(unreachable_pub)]
+pub use router_method::{
+    connect, delete, get, head, options, patch, post, put, trace, RouteMethod,
 };
-pub(crate) use tree::PathParams;