+++ poem/src/route/router.rs
@@ -3,16 +3,17 @@ use std::{str::FromStr, sync::Arc};
 use http::StatusCode;
 use regex::Regex;
 
-use super::tree::Tree;
 use crate::{
-    http::{uri::PathAndQuery, Method, Uri},
-    Body, Endpoint, EndpointExt, IntoEndpoint, IntoResponse, Request, Response,
+    endpoint::BoxEndpoint,
+    http::{uri::PathAndQuery, Uri},
+    route::internal::radix_tree::RadixTree,
+    Endpoint, EndpointExt, IntoEndpoint, IntoResponse, Request, Response,
 };
 
 /// Routing object
 #[derive(Default)]
 pub struct Route {
-    router: Tree<Box<dyn Endpoint<Output = Response>>>,
+    tree: RadixTree<BoxEndpoint<'static, Response>>,
 }
 
 impl Route {
@@ -104,7 +105,7 @@ impl Route {
         E: IntoEndpoint,
         E::Endpoint: 'static,
     {
-        self.router.add(
+        self.tree.add(
             &normalize_path(path.as_ref()),
             Box::new(ep.into_endpoint().map_to_response()),
         );
@@ -243,7 +244,7 @@ impl Route {
             false => 0,
             true => path.len() - 1,
         };
-        self.router.add(
+        self.tree.add(
             &format!("{}*--poem-rest", path),
             Box::new(Nest {
                 inner: ep.clone(),
@@ -251,7 +252,7 @@ impl Route {
                 prefix_len,
             }),
         );
-        self.router.add(
+        self.tree.add(
             &path[..path.len() - 1],
             Box::new(Nest {
                 inner: ep,
@@ -269,7 +270,7 @@ impl Endpoint for Route {
     type Output = Response;
 
     async fn call(&self, mut req: Request) -> Self::Output {
-        match self.router.matches(req.uri().path()) {
+        match self.tree.matches(req.uri().path()) {
             Some(matches) => {
                 req.state_mut().match_params.extend(matches.params);
                 matches.data.call(req).await
@@ -277,253 +278,6 @@ impl Endpoint for Route {
             None => StatusCode::NOT_FOUND.into(),
         }
     }
-}
-
-/// Routing object for HTTP methods
-#[derive(Default)]
-pub struct RouteMethod {
-    methods: Vec<(Method, Box<dyn Endpoint<Output = Response>>)>,
-}
-
-impl RouteMethod {
-    /// Create a `RouteMethod` object.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use poem::{
-    ///     handler,
-    ///     http::{Method, StatusCode},
-    ///     Endpoint, Request, RouteMethod,
-    /// };
-    ///
-    /// #[handler]
-    /// fn handle_get() -> &'static str {
-    ///     "get"
-    /// }
-    ///
-    /// #[handler]
-    /// fn handle_post() -> &'static str {
-    ///     "post"
-    /// }
-    ///
-    /// # tokio::runtime::Runtime::new().unwrap().block_on(async {
-    /// let route_method = RouteMethod::new().get(handle_get).post(handle_post);
-    ///
-    /// let resp = route_method
-    ///     .call(Request::builder().method(Method::GET).finish())
-    ///     .await;
-    /// assert_eq!(resp.status(), StatusCode::OK);
-    /// assert_eq!(resp.into_body().into_string().await.unwrap(), "get");
-    ///
-    /// let resp = route_method
-    ///     .call(Request::builder().method(Method::POST).finish())
-    ///     .await;
-    /// assert_eq!(resp.status(), StatusCode::OK);
-    /// assert_eq!(resp.into_body().into_string().await.unwrap(), "post");
-    /// # });
-    /// ```
-    pub fn new() -> Self {
-        Default::default()
-    }
-
-    /// Sets the endpoint for specified `method`.
-    pub fn method<E>(mut self, method: Method, ep: E) -> Self
-    where
-        E: IntoEndpoint,
-        E::Endpoint: 'static,
-    {
-        self.methods
-            .push((method, Box::new(ep.into_endpoint().map_to_response())));
-        self
-    }
-
-    /// Sets the endpoint for `GET`.
-    pub fn get<E>(self, ep: E) -> Self
-    where
-        E: IntoEndpoint,
-        E::Endpoint: 'static,
-    {
-        self.method(Method::GET, ep)
-    }
-
-    /// Sets the endpoint for `POST`.
-    pub fn post<E>(self, ep: E) -> Self
-    where
-        E: IntoEndpoint,
-        E::Endpoint: 'static,
-    {
-        self.method(Method::POST, ep)
-    }
-
-    /// Sets the endpoint for `PUT`.
-    pub fn put<E>(self, ep: E) -> Self
-    where
-        E: IntoEndpoint,
-        E::Endpoint: 'static,
-    {
-        self.method(Method::PUT, ep)
-    }
-
-    /// Sets the endpoint for `DELETE`.
-    pub fn delete<E>(self, ep: E) -> Self
-    where
-        E: IntoEndpoint,
-        E::Endpoint: 'static,
-    {
-        self.method(Method::DELETE, ep)
-    }
-
-    /// Sets the endpoint for `HEAD`.
-    pub fn head<E>(self, ep: E) -> Self
-    where
-        E: IntoEndpoint,
-        E::Endpoint: 'static,
-    {
-        self.method(Method::HEAD, ep)
-    }
-
-    /// Sets the endpoint for `OPTIONS`.
-    pub fn options<E>(self, ep: E) -> Self
-    where
-        E: IntoEndpoint,
-        E::Endpoint: 'static,
-    {
-        self.method(Method::OPTIONS, ep)
-    }
-
-    /// Sets the endpoint for `CONNECT`.
-    pub fn connect<E>(self, ep: E) -> Self
-    where
-        E: IntoEndpoint,
-        E::Endpoint: 'static,
-    {
-        self.method(Method::CONNECT, ep)
-    }
-
-    /// Sets the endpoint for `PATCH`.
-    pub fn patch<E>(self, ep: E) -> Self
-    where
-        E: IntoEndpoint,
-        E::Endpoint: 'static,
-    {
-        self.method(Method::PATCH, ep)
-    }
-
-    /// Sets the endpoint for `TRACE`.
-    pub fn trace<E>(self, ep: E) -> Self
-    where
-        E: IntoEndpoint,
-        E::Endpoint: 'static,
-    {
-        self.method(Method::TRACE, ep)
-    }
-}
-
-#[async_trait::async_trait]
-impl Endpoint for RouteMethod {
-    type Output = Response;
-
-    async fn call(&self, mut req: Request) -> Self::Output {
-        match self
-            .methods
-            .iter()
-            .find(|(method, _)| method == req.method())
-            .map(|(_, ep)| ep)
-        {
-            Some(ep) => ep.call(req).await,
-            None => {
-                if req.method() == Method::HEAD {
-                    req.set_method(Method::GET);
-                    let mut resp = self.call(req).await;
-                    resp.set_body(Body::empty());
-                    return resp;
-                }
-                StatusCode::NOT_FOUND.into()
-            }
-        }
-    }
-}
-
-/// A helper function, similar to `RouteMethod::new().get(ep)`.
-pub fn get<E>(ep: E) -> RouteMethod
-where
-    E: IntoEndpoint,
-    E::Endpoint: 'static,
-{
-    RouteMethod::new().get(ep)
-}
-
-/// A helper function, similar to `RouteMethod::new().post(ep)`.
-pub fn post<E>(ep: E) -> RouteMethod
-where
-    E: IntoEndpoint,
-    E::Endpoint: 'static,
-{
-    RouteMethod::new().post(ep)
-}
-
-/// A helper function, similar to `RouteMethod::new().put(ep)`.
-pub fn put<E>(ep: E) -> RouteMethod
-where
-    E: IntoEndpoint,
-    E::Endpoint: 'static,
-{
-    RouteMethod::new().put(ep)
-}
-
-/// A helper function, similar to `RouteMethod::new().delete(ep)`.
-pub fn delete<E>(ep: E) -> RouteMethod
-where
-    E: IntoEndpoint,
-    E::Endpoint: 'static,
-{
-    RouteMethod::new().delete(ep)
-}
-
-/// A helper function, similar to `RouteMethod::new().head(ep)`.
-pub fn head<E>(ep: E) -> RouteMethod
-where
-    E: IntoEndpoint,
-    E::Endpoint: 'static,
-{
-    RouteMethod::new().head(ep)
-}
-
-/// A helper function, similar to `RouteMethod::new().options(ep)`.
-pub fn options<E>(ep: E) -> RouteMethod
-where
-    E: IntoEndpoint,
-    E::Endpoint: 'static,
-{
-    RouteMethod::new().options(ep)
-}
-
-/// A helper function, similar to `RouteMethod::new().connect(ep)`.
-pub fn connect<E>(ep: E) -> RouteMethod
-where
-    E: IntoEndpoint,
-    E::Endpoint: 'static,
-{
-    RouteMethod::new().connect(ep)
-}
-
-/// A helper function, similar to `RouteMethod::new().patch(ep)`.
-pub fn patch<E>(ep: E) -> RouteMethod
-where
-    E: IntoEndpoint,
-    E::Endpoint: 'static,
-{
-    RouteMethod::new().patch(ep)
-}
-
-/// A helper function, similar to `RouteMethod::new().trace(ep)`.
-pub fn trace<E>(ep: E) -> RouteMethod
-where
-    E: IntoEndpoint,
-    E::Endpoint: 'static,
-{
-    RouteMethod::new().trace(ep)
 }
 
 fn normalize_path(path: &str) -> String {
@@ -645,72 +399,5 @@ mod tests {
         );
         assert_eq!(get(&r, "/a").await, "/");
         assert_eq!(get(&r, "/a?a=1").await, "/?a=1");
-    }
-
-    #[tokio::test]
-    async fn route_method() {
-        #[handler(internal)]
-        fn index() -> &'static str {
-            "hello"
-        }
-
-        for method in &[
-            Method::GET,
-            Method::POST,
-            Method::DELETE,
-            Method::PUT,
-            Method::HEAD,
-            Method::OPTIONS,
-            Method::CONNECT,
-            Method::PATCH,
-            Method::TRACE,
-        ] {
-            let route = RouteMethod::new().method(method.clone(), index).post(index);
-            let resp = route
-                .call(Request::builder().method(method.clone()).finish())
-                .await;
-            assert_eq!(resp.status(), StatusCode::OK);
-            assert_eq!(resp.into_body().into_string().await.unwrap(), "hello");
-        }
-
-        macro_rules! test_method {
-            ($(($id:ident, $method:ident)),*) => {
-                $(
-                let route = RouteMethod::new().$id(index).post(index);
-                let resp = route
-                    .call(Request::builder().method(Method::$method).finish())
-                    .await;
-                assert_eq!(resp.status(), StatusCode::OK);
-                assert_eq!(resp.into_body().into_string().await.unwrap(), "hello");
-                )*
-            };
-        }
-
-        test_method!(
-            (get, GET),
-            (post, POST),
-            (delete, DELETE),
-            (put, PUT),
-            (head, HEAD),
-            (options, OPTIONS),
-            (connect, CONNECT),
-            (patch, PATCH),
-            (trace, TRACE)
-        );
-    }
-
-    #[tokio::test]
-    async fn head_method() {
-        #[handler(internal)]
-        fn index() -> &'static str {
-            "hello"
-        }
-
-        let route = RouteMethod::new().get(index);
-        let resp = route
-            .call(Request::builder().method(Method::HEAD).finish())
-            .await;
-        assert_eq!(resp.status(), StatusCode::OK);
-        assert!(resp.into_body().into_vec().await.unwrap().is_empty());
     }
 }